% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{FR_recommendation}
\alias{FR_recommendation}
\title{This function computes the fertilizer estimate}
\usage{
FR_recommendation(lat, long, IntendedPlantingDate, country, rootUP,
  investment)
}
\arguments{
\item{lat}{Lattitude}

\item{long}{Longitude}

\item{IntendedPlantingDate}{controls not to ask recommendation for dates that are not considerd as planing dates in the zone}

\item{country}{Country e.g Nigeria, Tanzania}

\item{rootUP}{root rpice for fresh wt. ton/ha it is 270 in TZ and  141 for NG}

\item{investment:}{how much the farmer is willing to ivest  in USD (currently 200 USD)}
}
\value{
The estimate array
}
\description{
This function computes the fertilizer estimate
}
