% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizer.R
\name{nutrient_uptake}
\alias{nutrient_uptake}
\title{Nutrient uptake depends on the soil supply of the nutrient and the supply of other nutrients}
\usage{
nutrient_uptake(S1 = NA, S2 = NA, d1 = NA, a1 = NA, d2 = NA,
  a2 = NA, r1 = NA, r2 = NA)
}
\description{
Nutrient uptake depends on the soil supply of the nutrient and the supply of other nutrients
}
