% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizer.R
\name{optim_NR}
\alias{optim_NR}
\title{Optimize the UREA, TSP and MOP needed to maximize the NR. x1, x2, x3 = Urea, MOP and Nafaka kg/ha.}
\usage{
optim_NR(fertRate, rootUP, QID, CY, fertilizer, invest)
}
\description{
Optimize the UREA, TSP and MOP needed to maximize the NR. x1, x2, x3 = Urea, MOP and Nafaka kg/ha.
}
