% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{QUEFTS1_Pedotransfer}
\alias{QUEFTS1_Pedotransfer}
\title{computes target yield in tonnes/ha from a given NPK rate}
\usage{
QUEFTS1_Pedotransfer(QID, rec)
}
\arguments{
\item{QID}{a data frame containing soil NPK, WLY (kg/ha dry wt.),}

\item{rec}{recomended NPK rate}
}
\value{
target yield in ton/ha dry matter
}
\description{
computes target yield in tonnes/ha from a given NPK rate
}
\author{
Meklit
}
