% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizer.R
\name{recommendation}
\alias{recommendation}
\title{This function computes the fertilizer estimate}
\usage{
recommendation(lat, long, IntendedPlantingDate, countryObj, rootUP,
  investment)
}
\arguments{
\item{lat}{Lattitude}

\item{long}{Longitude}

\item{IntendedPlantingDate}{controls not to ask recommendation for dates that are not considerd as planing dates in the zone}

\item{countryObj}{Country code 1 is Nigeria 2 is Tanzania this is because the strings are not taken properly here and conversion os hectic}

\item{rootUP}{root rpice for fresh wt. ton/ha it is 270 in TZ and  141 for NG}

\item{investment:}{how much the farmer is willing to ivest  in USD (currently 200 USD)}
}
\value{
The estimate array
}
\description{
This function computes the fertilizer estimate
}
