% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizer.R
\name{Rerun_25kgKa_try}
\alias{Rerun_25kgKa_try}
\title{after setting fertilizer recommendation <25 kg/ha Urea, MOP or Nafaka, target yield with the remaining recommended fertilizer is  re-estimated  and
 total cost, gross and net revenue are re calcuated.}
\usage{
Rerun_25kgKa_try(rootUP, wdd, rdd, fertilizer, Quefts_Input_Data)
}
\arguments{
\item{rootUP}{cassava root price}

\item{wdd}{wdd}

\item{rdd}{rdd}

\item{fertilizer}{fertilizer type}

\item{zone}{zone}
}
\description{
after setting fertilizer recommendation <25 kg/ha Urea, MOP or Nafaka, target yield with the remaining recommended fertilizer is  re-estimated  and
 total cost, gross and net revenue are re calcuated.
}
\author{
Meklit
}
