% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{NPK_TargetYield_forOutput}
\alias{NPK_TargetYield_forOutput}
\title{using the output of function "NPK_TargetYield_forinput" and a dat frame per long and lat for intended NPK input
this function calculates the yield that can be obtained for intended NPK rate.}
\usage{
NPK_TargetYield_forOutput(NutrUse_soilNPK, N_rate, P_rate, K_rate)
}
\arguments{
\item{NPKdata:}{needs to be provided}
}
\value{

}
\description{
using the output of function "NPK_TargetYield_forinput" and a dat frame per long and lat for intended NPK input
this function calculates the yield that can be obtained for intended NPK rate.
}
\author{
Meklit
}
